adds isOwner middleware to check owner

+++ routes/location.js
@@ -2,6 +2,7 @@ const express = require('express');
 const api = express.Router();
 const db = require('../models/location.js');
 const auth = require('../middleware/auth');
+const isOwner = require('../middleware/isOwner');
 
 /**
  * get
@@ -9,6 +10,21 @@ const auth = require('../middleware/auth');
 api.get('/', async (req, res, next) => {
   try {
     const data = await db.find();
+    res.json(data);
+  } catch (err) {
+    next(err)
+    // throw new Error(err);
+  }
+
+})
+
+/**
+ * get
+ */
+api.get('/:id', async (req, res, next) => {
+  try {
+    const id = req.params.id;
+    const data = await db.findById(id);
     res.json(data);
   } catch (err) {
     next(err)
@@ -40,25 +56,36 @@ api.post('/', auth, async (req, res, next) => {
 /**
  * put
  */
-api.put('/:id', auth, async (req, res) => {
-  const id = typeof req.params.id === String ? req.params.id : null;
-  const updateCmd = {
-    $set: req.body
-  };
+api.put('/:id', auth, isOwner, async (req, res, next) => {
+  try {
+    
+    const id = req.params.id;
+    const updateCmd = {
+      $set: req.body
+    };
 
-  const updatedData = await db.updateOne(id, updateCmd);
-  res.json(updatedData);
+    const updatedData = await db.findByIdAndUpdate({_id:id}, updateCmd, {new:true});
+    res.json(updatedData);
+  } catch (err) {
+    next(err);
+  }
+
 })
 /**
  * delete
  */
 
-api.delete('/:id', auth, async (req, res) => {
-  const id = typeof req.params.id === String ? req.params.id : null;
-  await db.deleteOne(id);
-  res.json({
-    message: 'successfully deleted feature'
-  });
+api.delete('/:id', auth, isOwner, async (req, res, next) => {
+  try {
+    const id = typeof req.params.id === String ? req.params.id : null;
+    await db.deleteOne(id);
+    res.json({
+      message: 'successfully deleted feature'
+    });
+  } catch (error) {
+    next(err);
+  }
+
 })
